+++ src/Products/PluginIndexes/FieldIndex/FieldIndex.py
@@ -13,7 +13,7 @@
 
 from App.special_dtml import DTMLFile
 
-from Products.PluginIndexes.common.UnIndex import UnIndex
+from Products.PluginIndexes.unindex import UnIndex
 
 
 class FieldIndex(UnIndex):